+++ tests/functional/test_fuzz.py
@@ -269,13 +269,13 @@ class IteratorWrapper:
 class FuzzMachine(RuleBasedStateMachine):
     def __init__(self):
         super().__init__()
-        self.reinitialise()
+        self.reinitialise([])
 
-    def reinitialise(self):
-        self.key_type = None
-        self.sorted_keys = []
-        self.normal_dict = {}
-        self.sorted_dict = SortedDict()
+    def reinitialise(self, items):
+        self.key_type = type(items[0][0]) if items else None
+        self.sorted_keys = sorted(item[0] for item in items)
+        self.normal_dict = dict(items)
+        self.sorted_dict = SortedDict(items)
         self.sorted_dict_items = self.sorted_dict.items()
         self.sorted_dict_keys = self.sorted_dict.keys()
         self.sorted_dict_values = self.sorted_dict.values()
@@ -999,8 +999,22 @@ class FuzzMachine(RuleBasedStateMachine):
     ###########################################################################
 
     @rule()
-    def init(self):
-        self.reinitialise()
+    def init_wrong_call(self):
+        with pytest.raises(TypeError, match=re.escape("SortedDict() takes 0 to 1 positional arguments (2 given)")):
+            SortedDict(object, object)
+
+    @rule()
+    def init_not_iterable(self):
+        with pytest.raises(TypeError, match="object is not iterable"):
+            SortedDict(object)
+
+    @rule()
+    def init_empty(self):
+        self.reinitialise([])
+
+    @rule(good_other=rule_items_supported())
+    def init(self, good_other):
+        self.reinitialise(good_other)
 
 
 TestFuzz = FuzzMachine.TestCase