+++ docs/changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/tfpf/pysorteddict/compare/v0.14.0...main)
+
+### Changed
+
+* `SortedDict` initialiser inserts items from the first positional argument (if any)
+  ([#280](https://github.com/tfpf/pysorteddict/pull/280)).
+
 ## [0.14.0](https://github.com/tfpf/pysorteddict/compare/v0.13.1...v0.14.0) (2026-04-27)
 
 ### Added